+++ pan3d/ui/pan3d_components/module/__init__.py
@@ -1,7 +1,7 @@
 from pathlib import Path
 
 # Compute local path to serve
-serve_path = str(Path(__file__).parent.with_name("serve").resolve())
+serve_path = str(Path('pan3d/ui/pan3d_components/module'))
 
 # Serve directory for JS/CSS files
 serve = {"__pan3d_components": serve_path}
