+++ pan3d/ui/widgets.py
